+++ activemodel/CHANGELOG.md
@@ -1,2 +1,5 @@
+*   Add ActiveModel::Errors#excluding to limit attributes from list
+
+    *Nick Pezza*
 
 Please check [8-0-stable](https://github.com/rails/rails/blob/8-0-stable/activemodel/CHANGELOG.md) for previous changes.